requirements.txt
fastapi
uvicorn[Standard]
requests
boto3
python-jose
pydantic